+++ src/parser/PythonParser.hs
@@ -15,7 +15,7 @@ data Python = Class String
 
 instance Show Python where
   show (Class name) = "\"Class<s>" ++ name ++ "\""
-  show (Param name varType) = "Param<s>" ++ name ++ "<s>" ++ varType
+  show (Param name varType) = "Param<p>" ++ name ++ "<p>" ++ varType
   show (Function name params) = "\"Function<s>" ++ name ++ "<s>" ++ show params ++ "\""
   show (FunctionCall name) = "\"FunctionCall<s>" ++ name ++ "\""
   show (Property name varType value) = "\"Property<s>" ++ name ++ "<s>" ++ varType ++ "<s>" ++ value ++ "\""
@@ -24,15 +24,6 @@ instance Show Python where
   show (Comment comment) = "\"Comment<s>" ++ comment ++ "\""
 
 
--- parses variable names
-codeString :: Parser String
-codeString = spanP isCodeChar
-  where
-    isCodeChar c = if isAlpha c 
-                   then True 
-                   else case c of
-      '_' -> True
-      _   -> False
 
 -- parses called functions, i.e. variables names joined by .
 callString :: Parser String
@@ -46,16 +37,11 @@ callString = spanP isCallChar
       _   -> False
 
 
--- separates by parsing a delimiter and then parsing all elements and constructing a list
-sepBy :: Parser a -> Parser b -> Parser [b]
-sepBy sep element = (:) <$> element <*> many (sep *> element) <|> pure []  -- if fail, empty list
-
-
 -- parses a function parameter
 parseParam :: Parser Python
 parseParam = do
   name <- codeString <* ws
-  varType <- (charP ':' *> ws *> codeString <* ws) <|> pure ""
+  varType <- (charP ':' *> ws *> codeString <* ws) <|> pure "unset"
   Parser $ \input -> Just (input, Param name varType)